requirements.txt
# Core Framework - Fonctionnel
flask==2.3.3
flask-socketio==5.3.6
numpy<2.0.0
Pillow==10.0.1
qrcode==7.4.2
requests==2.31.0
geocoder==1.38.1
user-agents==2.2.0
psutil==5.9.6
python-socketio==5.9.0

# Interface et Utilitaires
colorama==0.4.6
rich==13.7.0
tqdm==4.66.1
pyfiglet==1.0.2
termcolor==1.1.0

# Fonctionnalités Avancées
selenium==4.15.2
beautifulsoup4==4.12.2
lxml==4.9.3
cryptography==41.0.7
pycryptodome==3.19.0

# Social Engineering
phonenumbers==8.13.25
email-validator==2.1.0
faker==20.1.0

# Network et DNS
dnspython==2.4.2
netaddr==0.8.0 